+++ packages/ui-link/src/Link/index.tsx
@@ -228,7 +228,6 @@ class Link extends Component<Props> {
       }
     })
 
-    // @ts-expect-error ts-migrate(2555) FIXME: Expected at least 5 arguments, but got 2.
     warn(
       // if display prop is used, warn about icon or TruncateText
       !truncateText || this.props.display === undefined,
@@ -253,12 +252,10 @@ class Link extends Component<Props> {
   }
 
   get interaction() {
-    // @ts-expect-error ts-migrate(2739) FIXME: Type 'Readonly<Props> & Readonly<{ children?: Reac... Remove this comment to see the full error message
     return getInteraction({ props: this.props, interactionTypes: ['disabled'] })
   }
 
   get element() {
-    // @ts-expect-error ts-migrate(2554) FIXME: Expected 3 arguments, but got 2.
     return getElementType(Link, this.props)
   }
 
@@ -282,7 +279,6 @@ class Link extends Component<Props> {
   }
 
   renderIcon() {
-    // @ts-expect-error ts-migrate(2555) FIXME: Expected at least 5 arguments, but got 2.
     warn(
       // if display prop is used, warn about icon or TruncateText
       this.props.display === undefined,