+++ packages/ui-options/src/Options/Item/index.tsx
@@ -83,10 +83,12 @@ class Item extends Component<Props> {
     role: PropTypes.string,
     /**
      * Content to render before the label
+     * (if you pass a function, it has the `props` as its parameter)
      */
     renderBeforeLabel: PropTypes.oneOfType([PropTypes.node, PropTypes.func]),
     /**
      * Content to render after the label
+     * (if you pass a function, it has the `props` as its parameter)
      */
     renderAfterLabel: PropTypes.oneOfType([PropTypes.node, PropTypes.func]),
     children: PropTypes.oneOfType([PropTypes.node, PropTypes.func])
@@ -114,7 +116,7 @@ class Item extends Component<Props> {
 
   // @ts-expect-error ts-migrate(7006) FIXME: Parameter 'renderLabel' implicitly has an 'any' ty... Remove this comment to see the full error message
   renderContent(renderLabel, contentVariant) {
-    const { styles } = this.props
+    const { styles, variant, as, role, children } = this.props
 
     return (
       <span
@@ -122,27 +124,24 @@ class Item extends Component<Props> {
         role="presentation"
         aria-hidden="true"
       >
-        {callRenderProp(renderLabel)}
+        {callRenderProp(renderLabel, {
+          variant,
+          as,
+          role,
+          children
+        })}
       </span>
     )
   }
 
   render() {
-    const {
-      as,
-      role,
-      styles,
-      renderBeforeLabel,
-      renderAfterLabel,
-      children
-    } = this.props
+    const { as, role, styles, renderBeforeLabel, renderAfterLabel, children } =
+      this.props
 
-    const ElementType = getElementType(Item, this.props, () => as)
-    // @ts-expect-error ts-migrate(2554) FIXME: Expected 3 arguments, but got 2.
+    const ElementType = getElementType(Item, this.props, () => as!)
     const passthroughProps = omitProps(this.props, Item.propTypes)
 
     return (
-      // @ts-expect-error ts-migrate(2322) FIXME: Type '{ children: (false | "" | 0 | Element | null... Remove this comment to see the full error message
       <ElementType role="none" css={styles.item}>
         <span {...passthroughProps} css={styles.container} role={role}>
           {children}