+++ packages/ui-pages/src/Pages/index.tsx
@@ -23,7 +23,7 @@
  */
 
 /** @jsx jsx */
-import React, { Component, createContext } from 'react'
+import React, { Component, createContext, ReactElement } from 'react'
 import PropTypes from 'prop-types'
 
 import { View } from '@instructure/ui-view'
@@ -200,11 +200,11 @@ class Pages extends Component<Props> {
     const activePage =
       activePageIndex < pages.length ? pages[activePageIndex] : null
 
-    // @ts-expect-error ts-migrate(2555) FIXME: Expected at least 5 arguments, but got 2.
+    // @ts-expect-error ts-migrate(2555) FIXME
     error(activePage, '[Pages] Invalid `activePageIndex`.')
 
     return activePage
-      ? safeCloneElement(activePage, {
+      ? safeCloneElement(activePage as ReactElement, {
           // @ts-expect-error ts-migrate(7006) FIXME: Parameter 'el' implicitly has an 'any' type.
           ref: (el) => {
             // @ts-expect-error ts-migrate(2551) FIXME: Property '_activePage' does not exist on type 'Pag... Remove this comment to see the full error message