+++ packages/ui-text/src/Text/index.tsx
@@ -33,7 +33,9 @@ import generateStyle from './styles'
 import generateComponentTheme from './theme'
 
 type Props = {
-  as?: React.ReactElement
+  //TODO: probably have to make a new type instead of this generic `string` which
+  //will only allow the valid html tags
+  as?: React.ReactElement | string
   color?:
     | 'primary'
     | 'secondary'
@@ -139,11 +141,9 @@ class Text extends Component<Props> {
   render() {
     const { children } = this.props
 
-    // @ts-expect-error ts-migrate(2554) FIXME: Expected 3 arguments, but got 2.
     const ElementType = getElementType(Text, this.props)
 
     return (
-      // @ts-expect-error ts-migrate(2322) FIXME: Type '{ children: ReactNode; css: any; ref: ((...a... Remove this comment to see the full error message
       <ElementType
         {...passthroughProps(this.props)}
         css={this.props.styles.text}