+++ packages/ui-buttons/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [8.6.0](https://github.com/instructure/instructure-ui/compare/v8.5.0...v8.6.0) (2021-06-18)
+
+### Performance Improvements
+
+- **ui-buttons:** use span instead of Flex ([e019e94](https://github.com/instructure/instructure-ui/commit/e019e94c7c8bce8434f7b348b7d9fd8e96b76191))
+
+# [8.5.0](https://github.com/instructure/instructure-ui/compare/v8.4.0...v8.5.0) (2021-06-09)
+
+### Bug Fixes
+
+- **emotion:** add componentId to match with componenet name in docs ([cc255eb](https://github.com/instructure/instructure-ui/commit/cc255ebef149ffe19a2102a179c1e7765b633888))
+
 # [8.4.0](https://github.com/instructure/instructure-ui/compare/v8.3.0...v8.4.0) (2021-05-11)
 
 ### Bug Fixes